Fix decodePayloadLazy and move it back into decodePayload. Fix Jest tests
