+++ src/lib/utilities/__mocks__/$app/env.ts
@@ -1,0 +1,1 @@
+export const browser = jest.fn();
