+++ src/lib/utilities/decode-payload.ts
@@ -1,9 +1,30 @@
 import type { Payload } from '$types';
-import { dataConverterPort } from '$lib/stores/data-converter-config';
 import { dataConverterWebsocket } from '$lib/utilities/data-converter-websocket';
-import { get } from 'svelte/store';
 import { convertPayload } from '$lib/services/data-converter';
-import { decodePayload } from './decode-payload-lazy';
+import { dataConverterPort } from '$lib/stores/data-converter-config';
+import { get } from 'svelte/store';
+
+export function decodePayload(
+  payload: Payload,
+  // This could decode to any object. So we either use the payload object passed in or decode it
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+): Payload | Record<any, any> | string {
+  const encoding = window.atob(String(payload.metadata.encoding));
+  // Help users out with an english encoding
+  (payload.metadata.encodingDecoded as unknown as string) = encoding;
+
+  switch (encoding) {
+    case 'json/plain':
+    case 'json/protobuf':
+      try {
+        return JSON.parse(window.atob(String(payload.data)));
+      } catch (_e) {
+        // Couldn't correctly decode this just let the user deal with the data as is
+      }
+  }
+
+  return payload;
+}
 
 export const convertPayloadToJson = async (
   eventAttribute: EventAttribute,
@@ -17,27 +38,28 @@ export const convertPayloadToJson = async (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const anyAttributes = eventAttribute as any;
 
-  const port = get(dataConverterPort);
-
   const potentialPayload = (anyAttributes?.input?.payloads ??
     anyAttributes?.result?.payloads ??
     null) as Payload[] | null;
 
-  if (potentialPayload === null) {
-    return Promise.resolve(eventAttribute);
-  }
+  // Decode payload data
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  let JSONPayload: string | Payload | Record<any, any>;
 
-  let JSONPayload = potentialPayload.map(decodePayload);
+  const port = get(dataConverterPort);
 
   if (port) {
+    // Convert Payload data
     const webSocket = dataConverterWebsocket.websocket;
 
     const awaitData = await Promise.all(
-      potentialPayload.map(
+      (potentialPayload ?? []).map(
         async (payload) => await convertPayload(payload, webSocket),
       ),
     );
     JSONPayload = awaitData;
+  } else {
+    JSONPayload = potentialPayload.map(decodePayload);
   }
 
   if (anyAttributes?.input?.payloads) {
